+++ src/core.c
@@ -266,6 +266,13 @@ SEXP rawToChar(const unsigned char *buf, const size_t sz) {
 
 void nano_serialize(nano_buf *buf, const SEXP object, SEXP hook) {
 
+  if (serial_alt) {
+    nano_qs2_loaded();
+    buf->buf = qs2_serialize(object, &buf->cur, 3, true, 1);
+    buf->len = buf->cur;
+    return;
+  }
+
   NANO_ALLOC(buf, NANONEXT_INIT_BUFSIZE);
   const int reg = hook != R_NilValue;
   int vec;
@@ -402,6 +409,9 @@ SEXP nano_unserialize(unsigned char *buf, const size_t sz, SEXP hook) {
       }
       cur = 12;
       goto resume;
+    case 0x0b:
+      nano_qs2_loaded();
+      return qs2_deserialize(buf, sz, false, 1);
     }
   }
 
@@ -435,6 +445,18 @@ SEXP nano_unserialize(unsigned char *buf, const size_t sz, SEXP hook) {
 
   return out;
 
+}
+
+void nano_qs2_loaded(void) {
+  if (qs2_deserialize != NULL)
+    return;
+  SEXP str, call;
+  PROTECT(str = Rf_mkString("qs2"));
+  PROTECT(call = Rf_lang2(Rf_install("loadNamespace"), str));
+  Rf_eval(call, R_BaseEnv);
+  UNPROTECT(2);
+  qs2_serialize = (unsigned char *(*)(SEXP, size_t *, const int, const bool, const int)) R_GetCCallable("qs2", "c_qs_serialize");
+  qs2_deserialize = (SEXP (*)(const unsigned char *, const size_t, const bool, const int)) R_GetCCallable("qs2", "c_qs_deserialize");
 }
 
 SEXP nano_decode(unsigned char *buf, const size_t sz, const uint8_t mod, SEXP hook) {
@@ -510,7 +532,7 @@ SEXP nano_decode(unsigned char *buf, const size_t sz, const uint8_t mod, SEXP ho
     return data;
   default:
     data = nano_unserialize(buf, sz, hook);
-  return data;
+    return data;
   }
 
   memcpy(NANO_DATAPTR(data), buf, sz);
@@ -658,5 +680,12 @@ int nano_matchargs(const SEXP mode) {
 SEXP rnng_eval_safe(SEXP arg) {
 
   return R_ToplevelExec(nano_eval_safe, arg) ? nano_eval_res : Rf_allocVector(RAWSXP, 1);
+
+}
+
+SEXP rnng_use_qs2(void) {
+
+  serial_alt = 1;
+  return R_NilValue;
 
 }