+++ NEWS.md
@@ -1,5 +1,9 @@
 # nanonext (development version)
 
+#### New Features
+
+* `use_qs2()` switches serialization to use the qs2 package.
+
 #### Updates
 
 * `write_cert()` argument 'cn' now defaults to '127.0.0.1' instead of 'localhost'.